Fjerner overflødig konvertering mellom Int og Long for offset-verdier

+++ src/main/kotlin/no/nav/familie/ba/sak/integrasjoner/økonomi/UtbetalingsperiodeMal.kt
@@ -38,8 +38,8 @@ data class UtbetalingsperiodeMal(
      */
     fun lagPeriodeFraAndel(
         andel: AndelTilkjentYtelse,
-        periodeIdOffset: Int,
-        forrigePeriodeIdOffset: Int?,
+        periodeIdOffset: Long,
+        forrigePeriodeIdOffset: Long?,
         opphørKjedeFom: YearMonth? = null
     ): Utbetalingsperiode =
         Utbetalingsperiode(
